requirements.txt
pandas
numpy
scikit-learn
pyyaml
pydantic
pytest
black
ruff
python-dotenv
joblib
azure-eventhub
